requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas>=2.2.0
scikit-learn>=1.4.1
joblib>=1.3.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dateutil>=2.8.2
